Move DeepChangeChecker to separate file
