+++ src/realm/object-store/CMakeLists.txt
@@ -19,6 +19,7 @@ set(SOURCES
 
     impl/collection_change_builder.cpp
     impl/collection_notifier.cpp
+    impl/deep_change_checker.cpp
     impl/list_notifier.cpp
     impl/set_notifier.cpp
     impl/object_notifier.cpp
@@ -63,6 +64,7 @@ set(HEADERS
 
     impl/collection_change_builder.hpp
     impl/collection_notifier.hpp
+    impl/deep_change_checker.hpp
     impl/external_commit_helper.hpp
     impl/list_notifier.hpp
     impl/set_notifier.hpp
